Fix ends_with test to avoid floating point exceptions
With certain random values the test would try to take the modulo of a
number with 0. This fixes it, and uses explicit
uniform_int_distributions to make the ranges of random values clearer.

This also updates the starts_with test to do the same, although it was
not buggy in the same way.