Add when_all_vector sender adaptor
This is an adaptor like when_all, but specialized to take a vector of
senders and signal the continuation when all senders in the vector have
completed. It requires that the senders send a single type, or none at
all. If they send nothing, the sender returned by when_all_vector also
sends nothing.