Remove enabler in then set_value
Fixes #181

The problem is that `tag_invoke` of any tag instantiates the enable_if, before the function is excluded from the overload set. For generic lambdas the body has to be instantiated to get the return type which is a hard error.
To my understanding there is nothing to disambiguate in the `then::tag_invoke(set_value, ...)` and it can be removed (alternatively the enabler can be put in a trailing return type).