Add a new bypass scheduler to execute tasks inliine on a pika thread
Callbacks to sender set_value/error should not be triggered directly
on a polling thread as continuations will be called on the polling
thread. Instead the execution should be tranferred to a pika
thread/task.

Adding a transfer adaptor costs task creation, plus enqueuing/dequeuing
of the task, so the bypass scheduler supports inline execution, but
with the creation of a task wrapper around the callback so that it is
safe to suspend/resume etc and no raw scheduler/polling threads are
running user code.