Avoid false positives in thread_pool_scheduler test
A thread created with transfer is not guaranteed to have a different id
from the previous thread. A thread description may be created for the
new thread, but the actual thread may be reused from the previous
thread. This disables a few tests which assumed that the new thread
would have a unique id.