Add completion mode and pool_name controls to mpi:: to control continuation execution
In general mpi invocations happen inline where they are placed by the
user, but the continuations that are triggered by their completion
should be explicitly transferred to a task and not run directly
on the polling thread.

Several options have been added to pass mpi invocations and/or the
continuations onto tasks on the specified pool. This also involved a
cleanup of how the mpi pool is named when used to ensure that the
transfer_mpi code knows where to place tasks.

Also categorize/prioritize and cleanup mpi_debug messsages

Correctly handle no mpi pool setup