Use queue for ready requests, handle them at start+end of poll function
This mimics the way cuda events are handled, any ready events are pushed
onto a lockfree queue and any thread can invoke the callback before
attempting to take the lock and queury for new ready requests.

If the lock is taken succcessfully, then read events are placed on the
queue and cleared again at the end of the poll function.