Replace limiting executor with limiting scheduler
This is a more sender/receiver conforming version
of the task limiting mechanism.

A test is provided that spawns tasks and does not allow
more than N to exist at a given time on the chosen scheduler.
Multiple limiting schedulers may be used at the same time and they
manage their own limiting independently.