Fix waiting in sync_wait
Remove shortcut to check if lock can be avoided as it can lead to the shared state being freed
before the notifying thread has released the lock. The following can happen with the shortcut:

- thread 2 takes lock to notify condition variable
- thread 2 sets set_called to true
- thread 1 sees set_called == true
- thread 1 skips the wait and returns from sync_wait, releasing the shared state
- thread 2 releases the lock (with a write), which was stored in the shared state

By keeping always taking the lock we guarantee that the notifying thread will always release the
lock before the waiting thread returns from the wait.