Store last_worker_thread_num of thread_data in atomic
The worker thread number can be read and written unsynchronized e.g. when one thread waits on a
condition variable and suspends itself, and another thread then immediately tries to notify the
first thread, reading the last worker thread number. Since the read and write of the worker thread
number is done without a mutex there's a data race. This commit changes the worker thread number to
be stored in an atomic. However, we do accept stale values and don't care if the number is reordered
with respect to other variables, so all reads and writes on the worker thread number are done using memory_order_relaxed.