Protect access to stealing masks in local_queue_scheduler.hpp
The masks are written and read potentially concurrently, by the thread setup and by the work
stealing in the scheduling loop. Since the masks are bitmasks, and not simply arrays of independent
bools, reading and writing different indices of the masks still end up accessing the same underlying
bytes, leading to data races. This commit protects the writes to the masks with a mutex, and
synchronizes the exit of on_start_thread with a barrier. The latter allows work stealing to read the
masks without locks.