Disable stealing of allocated threads with thread sanitizer enabled
It appears that thread sanitizer is not safe to use with stacks that migrate between os threads. See
https://groups.google.com/g/boost-developers-archive/c/xNNhBnfhANA and
https://stackoverflow.com/questions/73358829/segfault-in-tsan-in-boost-ci-need-the-help-of-an-expert
for a case reported in Boost asio/beast. Stacktraces in pika tend to look like the following when
stealing is enabled:

```
)(pika::latch&), pika::util::detail::pack_c<unsigned long, 0ul>, std::reference_wrapper<pika::latch> > >::then_sender_type, std::allocator<int> >::ensure_started_sender_type::shared_state::set_predecessor_done (this=0x7b280000a640) at /home/mjs/src/pika/libs/pika/execution
/include/pika/execution/algorithms/ensure_started.hpp:26
```

Note, stealing of staged threads (thread descriptions) is not disabled, as they haven't allocated or
used a stack yet.