Add task latency benchmark
The benchmark measures the time it takes to schedule a task and sync_wait for the result. The task
itself does no work and so the time measured is pure overhead. One can optionally request to to use
a stackless thread with the --nostack flag. The benchmark will schedule the other task on the next
worker thread from the spawning worker thread. Running the benchmark with only one worker thread
will schedule the other task on the same worker thread.