Fix data race when polling for CUDA events
Reading the size of the event vector (through empty) is done without a lock, while pushes and erases
are done with the lock. Reading without a lock leads to a data race and thread sanitizer reports
this race. We separately keep track of the number of events in the vector in an atomic, so we use
that to check if the vector is empty. We do a relaxed read since the check can be approximate.