Always set thread mask even when the mask is empty
When --pika:bind=none is requested the thread mask is empty, but this is equivalent to using all PUs
on the system. In the case that the process itself has a thread mask already set that is not the
full mask, not setting the full mask explicitly would lead to the wrong thread bindings.