Clean up version compatibility checks
Mixing compilers is a reasonable thing to do, on the user's responsibility. Generally, checking for
the same versions used during build and use of pika is a task better handled by a package
manager (it's also more completely covered by a package manager, instead of the ad-hoc checks that
were used here).