Fix some bypass scheduler issues when using dla-future
The bypass needs to be added to the cuda polling, but this termporary
fix inserts and extra bypass before issuing mpi requests if the thread
id is invalid.

Also, if the bypass scheduler is invoked on a thread that is already a
task, it skips task creation and becomes a simple pass through
execution.