Call spack spec without an explicit spec in CI
Since the environment is already concretized, calling it without an explicit spec simply prints the
spec of anything that is in the environment. Calling it with an explicit spec will reconcretize. It
will concretize to the right thing, unless something very strange happens, since all the
dependencies are installed, but it takes significantly longer to print the spec with an explicit spec.