Remove non_receiver_6 and rename non_receiver_7 to non_receiver_6
One static_assert on non_receiver_6 was checking if a receiver with a set_error without
noexcept is a receiver or not. non_receiver_6 without noexcept on
set_error is recognized as a receiver when we add the SFINAE to the cpo
in receiver.hpp. If we remove this SFINAE, the static_assert inside the
CPO that the set_error impl should be noexcept is triggered (what we
expect). Trying with stdexec (adding the PIKA_STDEXEC_RECEIVER_CONCEPT
to the non_receiver_6, so that we test the noexceptness of set_error has
the same behaviour as non stdexec code. It's recognized as a receiver
even without the noexceptness.

Since the behaviours are similar both with stdexec and without, I'm
removing this specific test case for now. It could be interesting to
figure out why the SFINAE discards the candidate, or to add a stronger
condition to satisfy the is_receiver trait.