Make sure async_rw_mutex senders always attach a continuation
When a sender is released without connecting it to a receiver, it is easier to attach a continuation
for lifetime management. This way, the shared state from a sender is always released through a
continuation instead of having multiple possible ways of being released, which is harder to manage
correctly.